Clima-API/requirements.txt
python-dotenv
os
requests